inference/requirements.txt
numpy
nibabel
scipy
scikit-image
tqdm
pandas
antspyx